requirements.txt
amqp==5.1.1
asgiref==3.2.10
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.1
django-celery-beat==2.2.1
django-environ==0.8.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
kombu==5.2.4
librabbitmq==2.0.0
prompt-toolkit==3.0.29
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
